requirements.txt
Flask>=2.0
gunicorn>=23.0.0
numpy
Pillow
joblib
scikit-learn
